codestyle: 18/12/2025 ReadMe

+++ ControlLibrary/Controls/TextControls/TextEditor/Util/LookupTable.cs
@@ -92,7 +92,7 @@ namespace ControlLibrary.Controls.TextControl.TextEditor.Util
 				// insert word into the tree
 				for (int i = 0; i < keyword.Length; ++i)
 				{
-					int index = ((int)keyword[i]) % 256; // index of curchar
+					int index = keyword[i] % 256; // index of curchar
 					bool d = keyword[i] == '\\';
 
 					next = next[index];             // get node to this index
